requirements.txt
python-telegram-bot==21.7
google-generativeai==0.8.5
openpyxl==3.1.5
python-dotenv==1.0.0
